Convert images to tensors during loading

+++ train.py
@@ -104,7 +104,7 @@ def generate_fid_stat(absolute_file_directory, IMAGE_SIZE = 256, img_size = 32):
     for images in os.listdir(absolute_file_directory):
         img_tensor.append(transform(PIL.Image.open(os.path.join(absolute_file_directory, images)).convert("RGB")))
     
-    return np.array(img_tensor)
+    return np.array(img_tensor, dtype = np.float64)
 
 
 fid_stat = generate_fid_stat(os.path.join(YOLO_DIR, 'raw_dataset', 'train', 'images'))
